refactor(NetworkError): 각 케이스 객체별로 분리
